Make all linking code exposed to 3rd parties
