+++ crates/cef-ui-helper/Cargo.toml
@@ -2,19 +2,14 @@
 name = "cef-ui-helper"
 version = "0.1.0"
 edition = "2021"
-build = "build.rs"
 
 [features]
 sandbox = []
 default = ["sandbox"]
-
-[build-dependencies]
-anyhow = { workspace = true }
 
 [dependencies]
 anyhow = { workspace = true }
 tracing = { workspace = true }
 tracing-log = { workspace = true }
 tracing-subscriber = { workspace = true }
-log = { workspace = true }
 libloading = "0.8"
