+++ crates/cef-ui-simple/Cargo.toml
@@ -18,5 +18,4 @@ anyhow = { workspace = true }
 tracing = { workspace = true }
 tracing-log = { workspace = true }
 tracing-subscriber = { workspace = true }
-log = { workspace = true }
 winit = "0.29"
