--- crates/cef-ui-tools/src/bin/cef_artifacts.rs
@@ -1,8 +1,0 @@
-use anyhow::Result;
-use cef_ui_tools::cef_artifacts;
-
-fn main() -> Result<()> {
-    cef_artifacts()?;
-
-    Ok(())
-}
