--- crates/cef-ui-tools/src/bin/cef_build.rs
@@ -1,8 +1,0 @@
-use anyhow::Result;
-use cef_ui_tools::cef_build;
-
-fn main() -> Result<()> {
-    cef_build()?;
-
-    Ok(())
-}
