--- crates/cef-ui-tools/src/bin/cef_clean.rs
@@ -1,8 +1,0 @@
-use anyhow::Result;
-use cef_ui_tools::cef_clean;
-
-fn main() -> Result<()> {
-    cef_clean()?;
-
-    Ok(())
-}
