+++ crates/cef-ui-util/Cargo.toml
@@ -5,9 +5,15 @@ edition = "2021"
 
 [dependencies]
 anyhow = { workspace = true }
+tracing = { workspace = true }
+tracing-log = { workspace = true }
+tracing-subscriber = { workspace = true }
 url = "2.5"
 percent-encoding = "2.3"
 bzip2 = "0.4"
 reqwest = { version = "0.12", features = ["blocking"] }
 tar = "0.4"
 flate2 = "1.0"
+walkdir = "2.5"
+bindgen = "0.69"
+clap = { version = "4.0", features = ["derive"] }
